--- src/main/java/com/elyo/nbox/file/controller/FileRequest.java
@@ -1,4 +1,0 @@
-package com.elyo.nbox.file.controller;
-
-public class FileRequest {
-}
