--- src/main/java/com/elyo/nbox/file/controller/FileResponse.java
@@ -1,4 +1,0 @@
-package com.elyo.nbox.file.controller;
-
-public class FileResponse {
-}
