+++ src/main/java/com/elyo/nbox/file/controller/dto/FileRequest.java
@@ -1,0 +1,4 @@
+package com.elyo.nbox.file.controller.dto;
+
+public class FileRequest {
+}
