--- src/main/java/com/elyo/nbox/file/service/FileService.java
@@ -1,10 +1,0 @@
-package com.elyo.nbox.file.service;
-
-
-import com.elyo.nbox.domain.IDownload;
-import com.elyo.nbox.domain.IUpload;
-import org.springframework.stereotype.Service;
-
-@Service
-public class FileService implements IUpload , IDownload {
-}
