+++ src/test/java/com/elyo/nbox/file/service/defaultFileUploadServiceTest.java
@@ -1,0 +1,14 @@
+package com.elyo.nbox.file.service;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+
+class defaultFileUploadServiceTest {
+
+    @Test
+    void upload() {
+
+    }
+}
